Implement the versions directory setup checker.
This required setting up logging so I can cheat and use
log levels to implement quiet and verbose level output.
I'm using `simplelog` as a sink, which in turn uses `paris`
to do some simple html-to-ansi-color-codes things to add
a little bit of styling to console messages. There is also
a file that gets logged to that is not subject to the
versbosity options and does more traditional timestamped
logging without the colors.

Implemented the core functions for walking directories and
returning lists of relevant files.

Expanded the command-line args a bit to support `--verbose` and
`--quiet` as well as to optionally pass a game directory for
both the check and the run commands..